+++ Plain.Forms/TreeViewEx.cs
@@ -31,7 +31,9 @@ namespace Plain.Forms {
 
 		protected override void OnHandleCreated(EventArgs e) {
 			base.OnHandleCreated(e);
-			NativeMethods.SendMessage(base.Handle, NativeMethods.TVM_SETEXTENDEDSTYLE, NativeMethods.TVS_EX_DOUBLEBUFFER | NativeMethods.TVS_EX_NOINDENTSTATE | NativeMethods.TVS_EX_FADEINOUTEXPANDOS, NativeMethods.TVS_EX_DOUBLEBUFFER | NativeMethods.TVS_EX_NOINDENTSTATE | NativeMethods.TVS_EX_FADEINOUTEXPANDOS);
+			NativeMethods.SendMessage(base.Handle, NativeMethods.TVM_SETEXTENDEDSTYLE,
+			    (IntPtr)(NativeMethods.TVS_EX_DOUBLEBUFFER | NativeMethods.TVS_EX_NOINDENTSTATE | NativeMethods.TVS_EX_FADEINOUTEXPANDOS),
+                (IntPtr)(NativeMethods.TVS_EX_DOUBLEBUFFER | NativeMethods.TVS_EX_NOINDENTSTATE | NativeMethods.TVS_EX_FADEINOUTEXPANDOS));
 			NativeMethods.SetWindowTheme(base.Handle, "explorer", null);
 			base.Indent = 0;
 		}
